+++ Dockerfile
@@ -139,8 +139,8 @@ RUN chmod +x /usr/local/bin/doctor && \
     chmod +x /usr/local/bin/worker-mails && \
     chmod +x /usr/local/bin/worker-messaging && \
     chmod +x /usr/local/bin/worker-webhooks && \
-    chmod +x /usr/local/bin/migrate-usage && \
-    chmod +x /usr/local/bin/usage-inf-calc
+    chmod +x /usr/local/bin/usage-create-metrics && \
+    chmod +x /usr/local/bin/usage-update-metrics
 
 # Letsencrypt Permissions
 RUN mkdir -p /etc/letsencrypt/live/ && chmod -Rf 755 /etc/letsencrypt/live/
