--- bin/migrate-usage
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-php /usr/src/code/app/cli.php migrate-usage $@
