+++ bin/usage-create-metrics
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+php /usr/src/code/app/cli.php usage-create-metrics $@
