--- bin/usage-inf-calc
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-php /usr/src/code/app/cli.php usage-inf-calc $@
