+++ bin/usage-update-metrics
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+php /usr/src/code/app/cli.php usage-update-metrics $@
