+++ src/Appwrite/Platform/Services/Tasks.php
@@ -2,8 +2,8 @@
 
 namespace Appwrite\Platform\Services;
 
-use Appwrite\Platform\Tasks\MigrateUsage;
-use Appwrite\Platform\Tasks\UsageInfCalc;
+use Appwrite\Platform\Tasks\UsageUpdateMetrics;
+use Appwrite\Platform\Tasks\UsageCreateMetrics;
 use Utopia\Platform\Service;
 use Appwrite\Platform\Tasks\Doctor;
 use Appwrite\Platform\Tasks\Install;
@@ -40,7 +40,7 @@ class Tasks extends Service
             ->addAction(SDKs::getName(), new SDKs())
             ->addAction(VolumeSync::getName(), new VolumeSync())
             ->addAction(Specs::getName(), new Specs())
-            ->addAction(MigrateUsage::getName(), new MigrateUsage())
-            ->addAction(UsageInfCalc::getName(), new UsageInfCalc());
+            ->addAction(UsageUpdateMetrics::getName(), new UsageUpdateMetrics())
+            ->addAction(UsageCreateMetrics::getName(), new UsageCreateMetrics());
     }
 }
